NIFI-5147 Moved HashAttribute to HashAttributeLegacy. Moved TestHashAttribute to TestHashAttributeLegacy. Moved CalculateAttributeHash to HashAttribute. Moved CalculateAttributeHashTest to HashAttributeTest. Updated processor manifest. Updated Javadoc.
