+++ nifi-nar-bundles/nifi-standard-bundle/nifi-standard-processors/src/test/java/org/apache/nifi/processors/standard/TestHashAttributeLegacy.java
@@ -16,7 +16,6 @@
  */
 package org.apache.nifi.processors.standard;
 
-import org.apache.nifi.processors.standard.HashAttribute;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.fail;
@@ -24,19 +23,17 @@ import static org.junit.Assert.fail;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-
 import org.apache.nifi.util.MockFlowFile;
 import org.apache.nifi.util.TestRunner;
 import org.apache.nifi.util.TestRunners;
-
 import org.junit.Test;
 
-public class TestHashAttribute {
+public class TestHashAttributeLegacy {
 
     @Test
     public void test() {
-        final TestRunner runner = TestRunners.newTestRunner(new HashAttribute());
-        runner.setProperty(HashAttribute.HASH_VALUE_ATTRIBUTE.getName(), "hashValue");
+        final TestRunner runner = TestRunners.newTestRunner(new HashAttributeLegacy());
+        runner.setProperty(HashAttributeLegacy.HASH_VALUE_ATTRIBUTE.getName(), "hashValue");
         runner.setProperty("MDKey1", ".*");
         runner.setProperty("MDKey2", "(.).*");
 
@@ -63,10 +60,10 @@ public class TestHashAttribute {
 
         runner.run(5);
 
-        runner.assertTransferCount(HashAttribute.REL_FAILURE, 1);
-        runner.assertTransferCount(HashAttribute.REL_SUCCESS, 4);
+        runner.assertTransferCount(HashAttributeLegacy.REL_FAILURE, 1);
+        runner.assertTransferCount(HashAttributeLegacy.REL_SUCCESS, 4);
 
-        final List<MockFlowFile> success = runner.getFlowFilesForRelationship(HashAttribute.REL_SUCCESS);
+        final List<MockFlowFile> success = runner.getFlowFilesForRelationship(HashAttributeLegacy.REL_SUCCESS);
         final Map<String, Integer> correlationCount = new HashMap<>();
         for (final MockFlowFile flowFile : success) {
             final String correlationId = flowFile.getAttribute("hashValue");
